+++ index.html
@@ -252,6 +252,8 @@
 
 </body>
 <!-- <script src="js/function.js"></script> -->
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
+<script src="js/util.js"></script>
+<script src="js/data.js"></script>
 
 </html>
